Created error handling that utilizes http responses for Users

+++ backend/src/main/java/backend/service/UserService.java
@@ -37,7 +37,11 @@ public class UserService {
         }).orElse(null);
     }
 
-    public void deleteUser(Long id){
+    public boolean deleteUser(Long id){
+        if (!userRepository.existsById(id)){
+            return false;
+        }
         userRepository.deleteById(id);
+        return true;
     }   
 }
